refactor(loaders): merge identityProviderConfig into identityProviders
- Add providers Map to IdentityProvidersConfig interface
- Combine module config query and providers query in one loader
- Remove separate identityProviderConfig loader (duplicate SQL eliminated)
- Update oauth.ts to use identityProviders.providers
