+++ packages/express-context/src/loaders/index.ts
@@ -14,8 +14,7 @@
  *   - authSettings    (metaschema_modules_public.sessions_module → tenant DB)
  *   - encryptedSecrets (metaschema_modules_public.config_secrets_user_module → user_secrets)
  *   - userAuth        (metaschema_modules_public.user_auth_module)
- *   - identityProviders (metaschema_modules_public.identity_providers_module)
- *   - identityProviderConfig (all enabled providers with decrypted secrets)
+ *   - identityProviders (metaschema_modules_public.identity_providers_module + providers Map)
  *   - connectedAccounts (metaschema_modules_public.connected_accounts_module)
  *
  * To add a new per-db lookup, implement a ModuleLoader and register it:
@@ -55,7 +54,6 @@ export { agentChatLoader } from './agent-chat';
 export { encryptedSecretsLoader } from './encrypted-secrets';
 export { userAuthLoader } from './user-auth';
 export { identityProvidersLoader } from './identity-providers';
-export { identityProviderConfigLoader } from './identity-provider-config';
 export { connectedAccountsLoader } from './connected-accounts';
 
 /**
@@ -74,7 +72,6 @@ import { agentChatLoader } from './agent-chat';
 import { encryptedSecretsLoader } from './encrypted-secrets';
 import { userAuthLoader } from './user-auth';
 import { identityProvidersLoader } from './identity-providers';
-import { identityProviderConfigLoader } from './identity-provider-config';
 import { connectedAccountsLoader } from './connected-accounts';
 
 export function createDefaultRegistry() {
@@ -91,7 +88,6 @@ export function createDefaultRegistry() {
   registry.register(encryptedSecretsLoader);
   registry.register(userAuthLoader);
   registry.register(identityProvidersLoader);
-  registry.register(identityProviderConfigLoader);
   registry.register(connectedAccountsLoader);
   return registry;
 }
